# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Подсистема 'Каршеринг'",
        "contact": {
            "email": "carsh-app@example.com"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://192.168.0.35:80",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/carsmodels": {
            "get": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Получение списка моделей машин",
                "description": "Display a listing of the resource.",
                "operationId": "/api/carsmodels(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список моделей машин"
                    }
                }
            },
            "post": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Добавить модель машины",
                "description": "Store a newly created resource in storage.",
                "operationId": "/api/carsmodels(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsModelsStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Добавление новой модели машиной"
                    }
                }
            }
        },
        "/api/carsmodels/2": {
            "get": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Вернуть конкретную модель машины",
                "description": "Display the specified resource.",
                "operationId": "/api/carsmodels/{carsmodels}(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает конкретную модель машины"
                    }
                }
            },
            "put": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Редактировать конкретную модель машины",
                "description": "Update the specified resource in storage.",
                "operationId": "/api/carsmodels/{carsmodels}(PUT)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsModelsStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирует конкретную модель машины"
                    }
                }
            }
        },
...
        "/api/carsoperations": {
            "get": {
                "tags": [
                    "Операции с машинами"
                ],
                "summary": "Получение списка операций по машинам",
                "description": "Display a listing of the resource.",
                "operationId": "/api/carsoperations(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список операций по машинам"
                    }
                }
            },
            "post": {
                "tags": [
                    "Операции с машинами"
                ],
                "summary": "Добавить операцию с машиной",
                "description": "Store a newly created resource in storage.",
                "operationId": "/api/carsoperations(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsOperationsStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создание новой операции с машиной"
                    }
                }
            }
        },
        "/api/carsoperations/2": {
            "get": {
                "tags": [
                    "Операции с машинами"
                ],
                "summary": "Вернуть конкретную операцию с машиной",
                "description": "Display the specified resource.",
                "operationId": "/api/carsoperations/{carsoperations}(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает конкретную операцию с машиной"
                    }
                }
            },
            "put": {
                "tags": [
                    "Операции с машинами"
                ],
                "summary": "Редактировать конкретную операцию с машиной",
                "description": "Update the specified resource in storage.",
                "operationId": "/api/carsoperations/{carsoperations}(PUT)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsOperationsStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирует конкретную операцию с машиной"
                    }
                }
            }
        },
...
        "/api/carsstatuses": {
            "get": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Вывести список статусов машин",
                "description": "Display a listing of the resource.",
                "operationId": "/api/carsstatuses(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список статусов машин"
                    }
                }
            },
            "post": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Добавить новый статус машины",
                "description": "Store a newly created resource in storage.",
                "operationId": "/api/carsstatuses(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsStatusesStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создает новый статус машины"
                    }
                }
            }
        },
        "/api/carsstatuses/1": {
            "get": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Вернуть статус конкретной машины",
                "description": "Display the specified resource.",
                "operationId": "/api/carsstatuses/{carsstatuses}(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает статус конкретного пользователя"
                    }
                }
            },
            "put": {
                "tags": [
                    "Статусы машин"
                ],
                "summary": "Редактировать статус машины",
                "description": "Update the specified resource in storage.",
                "operationId": "/api/carsstatuses/{carsstatuses}(PUT)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarsStatusesStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирует статус конкретной машины"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarsModelsStoreRequest": {
                "required": [
                    "car_model"
                ],
                "properties": {
                    "car_model": {
                        "description": "тестовая модель машины",
                        "type": "string",
                        "example": "test model"
                    },
                    "cars_mark_id": {
                        "description": "id марки машины",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "CarsOperationsStoreRequest": {
                "required": [
                    "car_operations"
                ],
                "properties": {
                    "cars_id": {
                        "description": "id машины",
                        "type": "integer",
                        "example": "1"
                    },
                    "users_id": {
                        "description": "id юзера",
                        "type": "integer",
                        "example": "2"
                    },
                    "cars_status_id": {
                        "description": "id статуса машины",
                        "type": "integer",
                        "example": "1"
                    },
                    "data_time_operation": {
                        "description": "дата и время опреации",
                        "type": "string",
                        "format": "date",
                        "example": "12.01.2023"
                    },
                    "GPS_cars_latitude": {
                        "description": "gps широта",
                        "type": "numeric",
                        "example": "32434.12333"
                    },
                    "GPS_cars_longitude": {
                        "description": "gps долгота",
                        "type": "numeric",
                        "example": "45434.72333"
                    }
                },
                "type": "object"
            },
            "CarsStatusesStoreRequest": {
                "required": [
                    "car_status"
                ],
                "properties": {
                    "car_status": {
                        "description": "тестовый статус машины",
                        "type": "string",
                        "example": "test"
                    }
                },
                "type": "object"
            },
...
        }
    }
}